Move top-level code to LsCommand
